// NFT-Publishing-System-Frontend/src/app/models/user-model.ts
export interface User {
  id: number;
  username: string;
  password: string;
  ethAddress: string;


  // constructor(
  //   id: number,
  //   username: string,
  //   password: string,
  //   firstName: string,
  //   lastName: string,
  //   emailId: string,
  //   userRole: string
  // ) {
  //   this.id = id;
  //   this.username = username;
  //   this.password = password;
  //   this.firstName = firstName;
  //   this.lastName = lastName;
  //   this.emailId = emailId;
  //   this.userRole = userRole;
  // }
}